2_MQTT_SQL Bridge/requirements.txt
paho-mqtt==1.6.1
mariadb==1.1.5.post3